requirements.txt
# Root requirements — shared across all agents
# Each agent also has its own requirements.txt for per-venv installs

httpx==0.27.2
pydantic==2.10.3
python-dotenv==1.0.1
pyyaml==6.0.2

# Microsoft auth (calendar_sync, morning_digest, grocery_optimizer)
msal==1.31.0

# Google APIs (calendar_sync, morning_digest, commute_ping)
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Discord bot (discord_reminders)
discord.py==2.3.2

# Testing
pytest==8.3.3
